+++ generators/ruby/codegen/package.json
@@ -1,0 +1,50 @@
+{
+  "name": "@fern-api/ruby-codegen",
+  "version": "0.0.0",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/fern-api/fern.git",
+    "directory": "generators/ruby/codegen"
+  },
+  "sideEffects": false,
+  "type": "module",
+  "exports": {
+    ".": {
+      "development": "./src/index.ts",
+      "source": "./src/index.ts",
+      "types": "./lib/index.d.ts",
+      "import": "./lib/index.js",
+      "default": "./lib/index.js"
+    }
+  },
+  "main": "lib/index.js",
+  "source": "src/index.ts",
+  "types": "lib/index.d.ts",
+  "files": ["lib"],
+  "scripts": {
+    "clean": "rm -rf ./lib",
+    "compile": "tsc",
+    "compile:debug": "tsc --sourceMap",
+    "depcheck": "depcheck",
+    "test": "vitest --passWithNoTests --run",
+    "test:debug": "pnpm run test --inspect --no-file-parallelism",
+    "test:update": "vitest --passWithNoTests --run -u"
+  },
+  "dependencies": {
+    "@fern-api/fs-utils": "workspace:*",
+    "@fern-fern/generator-exec-sdk": "^0.0.1167",
+    "@fern-fern/ir-sdk": "^39",
+    "lodash-es": "^4.17.21",
+    "url-join": "^5.0.0",
+    "zod": "^3.22.3"
+  },
+  "devDependencies": {
+    "@fern-api/configs": "workspace:*",
+    "@types/lodash-es": "^4.17.12",
+    "@types/node": "18.15.3",
+    "@types/url-join": "^4.0.3",
+    "depcheck": "^1.4.7",
+    "typescript": "5.9.3",
+    "vitest": "^4.0.8"
+  }
+}
