+++ generators/ruby/sdk/package.json
@@ -1,0 +1,54 @@
+{
+  "name": "@fern-api/fern-ruby-sdk",
+  "version": "0.0.0",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/fern-api/fern.git",
+    "directory": "generators/ruby/sdk"
+  },
+  "sideEffects": false,
+  "type": "module",
+  "exports": {
+    ".": {
+      "development": "./src/index.ts",
+      "source": "./src/index.ts",
+      "types": "./lib/index.d.ts",
+      "import": "./lib/index.js",
+      "default": "./lib/index.js"
+    }
+  },
+  "main": "lib/index.js",
+  "source": "src/index.ts",
+  "types": "lib/index.d.ts",
+  "files": [
+    "lib"
+  ],
+  "scripts": {
+    "clean": "rm -rf ./lib && rm -rf ./dist",
+    "compile": "tsc",
+    "compile:debug": "tsc --sourceMap",
+    "depcheck": "depcheck",
+    "dist:cli": "node build.mjs",
+    "env:prod": "env-cmd -r .env-cmdrc.cjs -e prod",
+    "publish:cli": "turbo run dist:cli --filter . && npm publish ./dist",
+    "test": "vitest --passWithNoTests --run",
+    "test:debug": "pnpm run test --inspect --no-file-parallelism",
+    "test:update": "vitest --passWithNoTests --run -u"
+  },
+  "devDependencies": {
+    "@fern-api/base-generator": "workspace:*",
+    "@fern-api/configs": "workspace:*",
+    "@fern-api/fern-ruby-model": "workspace:*",
+    "@fern-api/fs-utils": "workspace:*",
+    "@fern-api/logging-execa": "workspace:*",
+    "@fern-api/ruby-codegen": "workspace:*",
+    "@fern-api/ruby-generator-cli": "workspace:*",
+    "@fern-fern/generator-exec-sdk": "^0.0.1167",
+    "@fern-fern/ir-sdk": "^39",
+    "@types/node": "18.15.3",
+    "depcheck": "^1.4.7",
+    "typescript": "5.9.3",
+    "vitest": "^4.0.8",
+    "zod": "^3.22.3"
+  }
+}
