+++ generators/typescript-mcp/base/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@fern-api/configs/tsconfig/main.json",
+  "compilerOptions": {
+    "outDir": "lib",
+    "rootDir": "src"
+  },
+  "include": ["./src/**/*"],
+  "exclude": ["./src/asIs/**/*"]
+}
