+++ generators/typescript-mcp/model/package.json
@@ -1,0 +1,49 @@
+{
+  "name": "@fern-api/typescript-mcp-model",
+  "version": "0.0.0",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/fern-api/fern.git",
+    "directory": "generators/typescript-mcp/model"
+  },
+  "sideEffects": false,
+  "type": "module",
+  "exports": {
+    ".": {
+      "development": "./src/index.ts",
+      "source": "./src/index.ts",
+      "types": "./lib/index.d.ts",
+      "import": "./lib/index.js",
+      "default": "./lib/index.js"
+    }
+  },
+  "main": "lib/index.js",
+  "source": "src/index.ts",
+  "types": "lib/index.d.ts",
+  "files": ["lib"],
+  "scripts": {
+    "clean": "rm -rf ./lib && rm -rf ./dist",
+    "compile": "tsc",
+    "compile:debug": "tsc --sourceMap",
+    "depcheck": "depcheck",
+    "dist:cli": "node build.mjs",
+    "dockerTagLatest": "docker build -f ./Dockerfile -t fernapi/fern-typescript-mcp-model:latest ../../..",
+    "podmanTagLatest": "podman build -f ./Dockerfile -t fernapi/fern-typescript-mcp-model:latest ../../..",
+    "publish:cli": "turbo run dist:cli --filter . && npm publish ./dist",
+    "test": "vitest --passWithNoTests --run",
+    "test:debug": "pnpm run test --inspect --no-file-parallelism",
+    "test:update": "vitest --passWithNoTests --run -u"
+  },
+  "devDependencies": {
+    "@fern-api/base-generator": "workspace:*",
+    "@fern-api/configs": "workspace:*",
+    "@fern-api/fs-utils": "workspace:*",
+    "@fern-api/typescript-ast": "workspace:*",
+    "@fern-api/typescript-mcp-base": "workspace:*",
+    "@fern-fern/ir-sdk": "^58.2.0",
+    "@types/node": "18.15.3",
+    "depcheck": "^1.4.7",
+    "typescript": "5.9.3",
+    "vitest": "^4.0.8"
+  }
+}
