+++ packages/generator-cli/src/replay/fernignore.ts
